content/canvas/test/test_size.attributes.parse.zero.html

<!DOCTYPE HTML>
<title>Canvas test: size.attributes.parse.zero</title>
<script src="/MochiKit/MochiKit.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="0" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>
var _deferred = false;

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

ok(canvas.width == 0, "canvas.width == 0");

if (!_deferred) SimpleTest.finish();
});
</script>
